--- app/src/main/java/com/ikseong/ucompass/create/component/viewmodel/CreateViewModel.kt
@@ -1,10 +1,0 @@
-package com.ikseong.ucompass.create.component.viewmodel
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class CreateViewModel @Inject constructor(): ViewModel(){
-
-}
