+++ app/src/main/java/com/ikseong/ucompass/create/navgraph/createNavGraph.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.ikseong.ucompass.create.component.viewmodel.CreateViewModel
+import com.ikseong.ucompass.create.viewmodel.CreateViewModel
 import com.ikseong.ucompass.create.screen.CreateRoomFinishRoute
 import com.ikseong.ucompass.create.screen.CreateRoomTitleRoute
 import com.ikseong.ucompass.navigation.Routes
