+++ app/src/main/java/com/ikseong/ucompass/create/viewmodel/CreateUiAction.kt
@@ -1,4 +1,4 @@
-package com.ikseong.ucompass.create.component.viewmodel
+package com.ikseong.ucompass.create.viewmodel
 
 sealed interface CreateUiAction {
     data object OnCreateClick: CreateUiAction
