+++ app/src/main/java/com/ikseong/ucompass/create/viewmodel/CreateViewModel.kt
@@ -1,0 +1,24 @@
+package com.ikseong.ucompass.create.viewmodel
+
+import android.content.Context
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import dagger.hilt.android.qualifiers.ApplicationContext
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import javax.inject.Inject
+
+@HiltViewModel
+class CreateViewModel @Inject constructor(
+    @ApplicationContext private val context: Context
+) : ViewModel() {
+
+    private val _uiState = MutableStateFlow(CreateUiState())
+    val uiState = _uiState.asStateFlow()
+
+    fun updateTitle(title: String) {
+        _uiState.value = _uiState.value.copy(title = title)
+    }
+
+}
+
